Move: clean up detached files to assets/images, screen/authentication
