+++ front/assets/images/authentication/loro_logo.png
