+++ front/lib/main.dart
@@ -3,9 +3,9 @@ import 'package:capstone/screen/authentication/setup_user.dart';
 import 'package:capstone/screen/authentication/social_login.dart';
 import 'package:capstone/constants/fonts.dart' as fonts;
 import 'package:capstone/screen/bottom_navigation.dart';
-import 'package:capstone/screen/sign_up/audio_player.dart';
-import 'package:capstone/screen/sign_up/controller/user_controller.dart';
-import 'package:capstone/screen/sign_up/get_user_voice.dart';
+import 'package:capstone/screen/authentication/audio_player.dart';
+import 'package:capstone/screen/authentication/controller/user_controller.dart';
+import 'package:capstone/screen/authentication/get_user_voice.dart';
 import 'package:capstone/widget/audio_recoder/recording_section.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
