+++ front/pubspec.yaml
@@ -80,10 +80,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - .env
-    - assets/icons/authentication/login/loro_logo.png
-    - assets/icons/authentication/login/google_icon.png
-    - assets/icons/authentication/login/tos_line.svg
-    - assets/icons/authentication/login/policy_line.svg
+    - assets/images/authentication/
     - assets/images/characters/
     - assets/wavs/
